change LatePenaltyCalculator to PenaltyCalculator
This will someday allow for a Scorer to call multiple PenaltyCalculators that decide if a submission should be blocked and which user would need to resolve it.
